+++ lib/src/handler/w3c_handler.dart
@@ -20,6 +20,7 @@ import 'package:appium_driver/src/handler/w3c/ime.dart';
 import 'package:appium_driver/src/handler/w3c/device.dart';
 import 'package:appium_driver/src/handler/w3c/settings.dart';
 import 'package:appium_driver/src/handler/w3c/sessions.dart';
+import 'package:appium_driver/src/handler/w3c/status.dart';
 import 'package:webdriver/src/common/request.dart'; // ignore: implementation_imports
 import 'package:webdriver/src/handler/w3c/utils.dart'; // ignore: implementation_imports
 
@@ -29,6 +30,9 @@ class W3cWebDriverHandler extends AppiumWebDriverHandler {
 
   @override
   final SessionsHandler sessions = W3cSessionsHandler();
+
+  @override
+  final StatusHandler status = W3cStatusHandler();
 
   @override
   final CoreHandler core = W3cCoreHandler();